vouchers can be set to live longer than just one registration

+++ src/inc/api/APIRegisterAgent.class.php
@@ -7,7 +7,8 @@ use DBA\RegVoucher;
 
 class APIRegisterAgent extends APIBasic {
   public function execute($QUERY = array()) {
-    global $FACTORIES;
+    /** @var $CONFIG DataSet */
+    global $FACTORIES, $CONFIG;
     
     //check required values
     if (!PQueryRegister::isValid($QUERY)) {
@@ -25,7 +26,10 @@ class APIRegisterAgent extends APIBasic {
     //create access token & save agent details
     $token = Util::randomString(10);
     $agent = new Agent(0, $name, "", -1, "", "", 0, 1, 0, $token, PActions::REGISTER, time(), Util::getIP(), null, 0, "");
-    $FACTORIES::getRegVoucherFactory()->delete($voucher);
+    
+    if ($CONFIG->getVal(DConfig::VOUCHER_DELETION) == 0) {
+      $FACTORIES::getRegVoucherFactory()->delete($voucher);
+    }
     $agent = $FACTORIES::getAgentFactory()->save($agent);
     if ($agent != null) {
       $payload = new DataSet(array(DPayloadKeys::AGENT => $agent));
